Solve empty field issue
