+++ packages/desmos-std/src/query_router.rs
@@ -10,6 +10,7 @@ use crate::{
             QueryUserGroupMembersResponse,
             QueryUserPermissionsResponse
         },
+        query_router::SubspacesQuerier,
         routes::SubspacesRoutes
     }
 };
@@ -23,15 +24,6 @@ impl<'a> DesmosQuerier<'a> {
     pub fn new(querier: &'a QuerierWrapper<'a, DesmosQueryRouter>) -> Self {
         DesmosQuerier { querier }
     }
-}
-
-pub trait SubspacesQuerier {
-    fn query_subspaces(&self, pagination: Option<PageRequest>) ->  StdResult<QuerySubspacesResponse>;
-    fn query_subspace(&self, subspace_id: u64) -> StdResult<QuerySubspaceResponse>;
-    fn query_user_groups(&self, subspace_id: u64, pagination: Option<PageRequest>) -> StdResult<QueryUserGroupsResponse>;
-    fn query_user_group(&self, subspace_id: u64, group_id : u32) -> StdResult<QueryUserGroupResponse>;
-    fn query_user_group_members(&self, subspace_id: u64, group_id : u32, pagination: Option<PageRequest>) -> StdResult<QueryUserGroupMembersResponse>;
-    fn query_user_permissions(&self, subspace_id: u64, user : String) -> StdResult<QueryUserPermissionsResponse>;
 }
 
 impl <'a> SubspacesQuerier for DesmosQuerier<'a> {
