+++ packages/desmos-std/src/subspaces/mod.rs
@@ -1,3 +1,4 @@
 pub mod models;
 pub mod query;
 pub mod routes;
+pub mod query_router;
